refactor stitching into module
