+++ mantis/cli/main.py
@@ -7,6 +7,7 @@ from mantis.cli.estimate_bleaching import estimate_bleaching
 from mantis.cli.estimate_deskew import estimate_deskew
 from mantis.cli.optimize_affine import optimize_affine
 from mantis.cli.run_acquisition import run_acquisition
+from mantis.cli.stitch import stitch_zarr_store
 from mantis.cli.update_scale_metadata import update_scale_metadata
 
 CONTEXT = {"help_option_names": ["-h", "--help"]}
@@ -31,3 +32,4 @@ cli.add_command(estimate_affine)
 cli.add_command(optimize_affine)
 cli.add_command(apply_affine)
 cli.add_command(update_scale_metadata)
+cli.add_command(stitch_zarr_store)
